BUG: Share underlying buffer from client/server to PDataReader
By creating a new read buffer in PDataReader, we potentially miss out on
any bytes that the `receive()` method read but didn't process yet.

Additionally, we lose the opposite when returning out of `PDataReader`:
remaining data in the PDataReader buffer that it didn't process.

The solution here is to mutably share that underlying read buffer with the
`PDataReader` so that it can continue from where the `receive()` method
left off _and_ pass any data that it didn't read back to the read buffer
of the main `Client` or `Server` object.