[core] Make `PersonName` hold both borrowed and owned strings
At the moment, `PersonName` cannot be ergonomically used by application
code as an structure to group and pass around owned person name
components that are meant to conform to DICOM's PN value semantics.
It is, for example, difficult to come up with code for deserializing a
`PersonName` if strings cannot be borrowed from the data source,
which is a common ocurrence when the source is e.g. a stream that reads
data chunks to temporary buffers.

To improve on that, let's make `PersonName` have `Cow<'a, str>` fields
instead, and update `PersonNameBuilder` to accept them. `Cow`s are smart
pointers that can hold both borrowed and owned data, so while `&'a str`
can be wrapped in a `Cow<'a, str>` at virtually no cost, an owned
`String` may also be wrapped with the same type but an indefinitely long
lifetime, `Cow<'static, str>`. Therefore, code that borrows name
components can still use `PersonName`s as before, and new code that
relies on `PersonName`s owning their components can now easily work by
just using a `'static` lifetime bound.