Add ServerAssociation::{acceptor,requestor}_max_pdu_length()
Modify association_echo.rs to verify that the results of the functions match the expectations, and to make both parts send and receive a packet with the maximum size admitted by each, with different sizes for the client and the server. The choice of which side gets the highest is random.