Return the abstract syntax in the presentation_contexts() method
This is a breaking change.

The presentation_contexts() method now returns a vector of a new PresentationContextNegotiated struct which includes all fields that were in PresentationContextResult, plus a new abstract_syntax String field, and the old PresentationContextResult stays for messaging only.

The client can no longer return presentation context entries sent by a broken server that include an ID that was not among the ones proposed.