Fix confusion between buffer size and variable field size
According to the standard PS3.8 Annex D tables D.1-1 and D.1-2 <https://dicom.nema.org/medical/dicom/2025c/output/chtml/part08/chapter_D.html>, the maximum PDU length that the peers communicate to each other during association establishment, applies to "the variable field of the P-DATA-TF PDUs". The "variable field" refers to the bytes starting on the 7th, according to PS3.8 section 9.3.5 <https://dicom.nema.org/medical/dicom/2025c/output/chtml/part08/sect_9.3.5.html>, and thus does not include the header. The code was interpreting it as a mix of the correct length (especially in reader.rs and writer.rs), the length including the PDU header (especially in server.rs and client.rs) and the length including the PDU and the PDV header (especially in pdata.rs).

This attempts to sort out the mess and uniformly use the same interpretation as the one the standard uses. A lot of it has to do with buffer size allocation: the buffers need room for maximum PDU size + the header.